requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==10.4.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.3
